refactor(vetting): Phase 3b-1 — DTO int → enum normalization
Second sub-commit of Phase 3 (backend vetting status normalization).
Phase 3b-1 converts every admin DTO's vettingStatus / WorkflowStatus
field from `int` to the `VettingStatus` enum type, so these DTOs now
ship their JSON values as enum strings (e.g. "UnderReview", "Approved")
via the globally-registered JsonStringEnumConverter.

The ApplicationUser entity itself still stores VettingStatus as `int`
— Phase 3b-2 will normalize the entity and remove the temporary
(VettingStatus)user.VettingStatus casts that this commit introduces
in DTO constructors and LINQ projections. Splitting the entity change
into a separate commit keeps the blast radius of Phase 3b-1 contained
to DTOs + their direct consumers.

## Backend changes

### DTOs (6 files, 7 fields)

- apps/api/Features/Users/Models/UserDto.cs
  VettingStatus: int → VettingStatus. Constructor casts from entity.

- apps/api/Features/Users/Models/UpdateUserRequest.cs
  VettingStatus: int? → VettingStatus?. Admin clients can now send
  enum strings for status updates.

- apps/api/Features/Users/Models/MemberDetails/MemberDetailsModels.cs
  MemberDetailsResponse.VettingStatus: int → VettingStatus
  VettingDetailsResponse.WorkflowStatus: int? → VettingStatus?

- apps/api/Features/EmailTemplates/Models/UserPreviewDto.cs
  VettingStatus: int → VettingStatus

- apps/api/Features/TestHelpers/Models/CreateTestUserRequest.cs
  VettingStatus: int → VettingStatus (default: VettingStatus.UnderReview)

- apps/api/Features/VettingHold/Models/VettingHoldModels.cs
  MembershipHoldResponse.NewStatus: int → VettingStatus
  VettingHoldStatusResponse.VettingStatus: int → VettingStatus

### Service consumers (5 files)

- UserManagementService.cs — cast `request.VettingStatus.Value` to int
  when assigning to entity (entity stays int for 3b-1).
- MemberDetailsService.cs — added Vetting.Entities using; switch cases
  now use named enum values instead of magic integers; DTO constructions
  cast entity int to enum; VettingDetailsResponse.WorkflowStatus
  assignment simplified (VettingApplication already uses the enum type,
  so the old `(int)application.WorkflowStatus` cast was unnecessary).
- EmailTemplateService.cs — added cast in UserPreviewDto Select
  projection (runs on materialized objects, not LINQ-to-SQL).
- TestHelperService.cs — cast when assigning to entity.
- VettingHoldService.cs — DTO construction sites now use enum literals
  (VettingStatus.OnHold, VettingStatus.FinalReview) instead of the
  private integer constants; cast for projecting entity int into the
  VettingHoldStatusResponse. The private integer constants remain
  because they're still used for entity comparisons until Phase 3b-2.

### Generated artifacts

- apps/api/openapi.json — regenerated from backend build
- packages/shared-types/src/generated/*.ts — regenerated via
  `npm run generate` in packages/shared-types. All admin DTOs now
  have `vettingStatus?: components["schemas"]["VettingStatus"]`
  instead of `vettingStatus?: number`.

## Frontend changes

- apps/web/src/pages/admin/AdminDashboardPage.tsx — switched from
  `isActiveMemberStatus(int)` to `isActiveMemberStatusString(enum)`
  helper. The active members count filter now consumes the enum string
  directly, eliminating the need for the int→string bridge lookup.

- apps/web/src/features/admin/members/components/MembersList.tsx —
  removed the `STATUS_INT_TO_STRING[user.vettingStatus]` lookup in
  getVettingStatusBadge; the switch now operates directly on the enum
  string. Labels and colors stay inline by design (documented in the
  file comment carried forward from Phase 2f — the admin members table
  intentionally uses different labels/colors than the applicant-facing
  config).

- apps/web/src/features/vetting/constants/vettingStatusConfig.ts —
  DELETED the Phase 2a temporary int bridge. Removed:
    * STATUS_INT_TO_STRING constant
    * getConfigFromInt(value: number) function
    * isActiveMemberStatus(value: number) function (int variant)
  Kept isActiveMemberStatusString as the sole active-member predicate.
  The header comment updated to explain what was removed and why.

## Test files (updated by test-executor agent)

The test-executor agent ran the full .NET test suite via the
run-test-suite skill and found 0 Phase 3b-1 regressions. 5 test files
required trivial updates to match the new DTO types (int literal
assertions/initializers → VettingStatus enum values):

- tests/unit/api/Features/VettingHold/VettingHoldEndpointsTests.cs
  (8 sites in MembershipHoldResponse and VettingHoldStatusResponse assertions)
- tests/unit/api/Features/Users/MemberDetailsEndpointsTests.cs
  (2 initializers)
- tests/unit/api/Features/Users/MemberDetailsServiceBusinessLogicTests.cs
  (3 assertions)
- tests/WitchCityRope.Core.Tests/Features/VettingHold/VettingHoldServiceTests.cs
  (4 DTO assertions; entity assignments kept as int — they'll update in Phase 3b-2)
- tests/integration/api/Features/VettingHold/VettingHoldIntegrationTests.cs
  (4 DTO assertions; entity assertions kept as int — same reason)

## Verification

- `dotnet build apps/api/WitchCityRope.Api.csproj`: 0 errors, 75
  pre-existing warnings
- `restart-dev-containers` skill: 3/3 containers healthy
- `npm run generate` in packages/shared-types: successful regen,
  tsc validation passed
- `tsc --noEmit` in apps/web: 0 errors in Phase 3b-1 files (2
  pre-existing EventForm.tsx errors remain — tech debt #1, unrelated)
- `run-test-suite` delegation to test-executor agent:
  * .NET Unit Tests: 1064/1110 pass, 35 pre-existing failures, 0 new regressions
  * .NET Core Tests: 106/132 pass, 8 pre-existing failures, 0 new regressions
  * .NET Integration: 172/267 pass, 84 pre-existing failures, 0 new regressions
  * All Phase 3b-1-scoped tests pass 100% (VettingHold endpoints 19/19,
    VettingHold core 24/24)
  * React vitest suite NOT run — no skill wrapper exists for vitest;
    verified via tsc clean in all Phase 3b-1 frontend files instead

## Phase 3b-2 preview

Phase 3b-2 will change ApplicationUser.VettingStatus from int to the
enum, which will cascade through 15+ service files (EmailTemplateService,
AdHocEmailSendJob, AttendanceService IsVetted checks, ProxyRsvpService,
VettingService, VettingSeeder, etc.) and remove every temporary cast
this commit introduced. It's a larger atomic change that needs its
own testing pass.

## Scope discipline

- Testing agent committed their work earlier (9a815064, a7e9d13a,
  e1c0dd8e). I'm staging only files I touched in Phase 3a + 3b-1 plus
  test files the test-executor agent updated to match my type changes.
- No manual test runner invocations; all tests ran via the
  run-test-suite skill.

+++ apps/api/Features/TestHelpers/Services/TestHelperService.cs
@@ -69,7 +69,9 @@ public class TestHelperService : ITestHelperService
                 FirstName = request.FirstName,
                 LastName = request.LastName,
                 Role = request.Role ?? "", // Empty string = no special role; "Member" is not a valid role
-                VettingStatus = request.VettingStatus, // 0-6 enum value, 3 = Approved (vetted)
+                // Phase 3b-1: request.VettingStatus is now the VettingStatus
+                // enum; the entity still stores int (Phase 3b-2 will align it).
+                VettingStatus = (int)request.VettingStatus,
                 Bio = request.Bio,
                 Pronouns = request.Pronouns ?? string.Empty, // CRITICAL: Database has NOT NULL constraint
                 DateOfBirth = dateOfBirth,
